+++ core-ng/src/main/java/core/framework/json/JSONException.java
@@ -7,6 +7,10 @@ public class JSONException extends RuntimeException {
     @Serial
     private static final long serialVersionUID = 9098858367692044462L;
 
+    public JSONException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
     public JSONException(Throwable cause) {
         super(cause);
     }